+++ src/test_scenes/test_scene_settings.h
@@ -3,6 +3,6 @@
 struct test_scene_settings {
 	// GEN INTROSPECTOR struct test_scene_settings
 	bool create_minimal = false;
-	float scene_tickrate = 144.f;
+	unsigned scene_tickrate = 144;
 	// END GEN INTROSPECTOR
 };
