+++ src/view/rendering_scripts/illuminated_rendering.cpp
@@ -171,7 +171,7 @@ void illuminated_rendering(
 			messages::visibility_information_request request;
 			request.eye_transform = *viewed_character_transform;
 			request.filter = predefined_queries::pathfinding();
-			request.square_side = std::max(screen_size.x, screen_size.y) * 2;
+			request.square_side = static_cast<float>(std::max(screen_size.x, screen_size.y) * 2);
 			request.subject = viewed_character;
 
 			auto responses = visibility_system(DEBUG_LOGIC_STEP_LINES).calc_visibility(
